-: Refactor FeedMeOfferStoreShake
